bug fix for Master Conjuror, Molten Core, Imp Glyph, and make Improved Divine Spirit not stack with other spell power buffs
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@689 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ simcraft.h
@@ -445,7 +445,7 @@ struct player_t
     double    flametongue_totem;
     player_t* focus_magic;
     int8_t    focus_magic_feedback;
-    int8_t    improved_divine_spirit;
+    double    improved_divine_spirit;
     int8_t    improved_moonkin_aura;
     int8_t    innervate;
     int8_t    lightning_capacitor;
